# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FC Rating API",
    "version": "0.1.0",
    "description": "Elo-based match rating API for a friend group's EA Sports FC league."
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "fc_session"
      }
    },
...
  },
  "paths": {
...
    "/matches/preview": {
      "post": {
        "operationId": "previewMatch",
        "summary": "Preview a match outcome without recording it (requires recorder role)",
        "tags": [
          "matches"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "homePlayerId": {
                    "type": "string"
                  },
                  "awayPlayerId": {
                    "type": "string"
                  },
                  "homeScore": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 99
                  },
                  "awayScore": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 99
                  }
                },
                "required": [
                  "homePlayerId",
                  "awayPlayerId",
                  "homeScore",
                  "awayScore"
                ]
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "home": {
                      "type": "object",
                      "properties": {
                        "playerId": {
                          "type": "string"
                        },
                        "before": {
                          "type": "object",
                          "properties": {
                            "rating": {
                              "type": "number"
                            },
                            "gamesPlayed": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "rating",
                            "gamesPlayed"
                          ],
                          "additionalProperties": false
                        },
                        "after": {
                          "type": "object",
                          "properties": {
                            "rating": {
                              "type": "number"
                            },
                            "gamesPlayed": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "rating",
                            "gamesPlayed"
                          ],
                          "additionalProperties": false
                        },
                        "expectedScore": {
                          "type": "number"
                        },
                        "actualScore": {
                          "anyOf": [
                            {
                              "type": "number",
                              "enum": [
                                1
                              ]
                            },
                            {
                              "type": "number",
                              "enum": [
                                0.5
                              ]
                            },
                            {
                              "type": "number",
                              "enum": [
                                0
                              ]
                            }
                          ]
                        },
                        "delta": {
                          "type": "number"
                        },
                        "wasProvisional": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "playerId",
                        "before",
                        "after",
                        "expectedScore",
                        "actualScore",
                        "delta",
                        "wasProvisional"
                      ],
                      "additionalProperties": false
                    },
                    "away": {
                      "type": "object",
                      "properties": {
                        "playerId": {
                          "type": "string"
                        },
                        "before": {
                          "type": "object",
                          "properties": {
                            "rating": {
                              "type": "number"
                            },
                            "gamesPlayed": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "rating",
                            "gamesPlayed"
                          ],
                          "additionalProperties": false
                        },
                        "after": {
                          "type": "object",
                          "properties": {
                            "rating": {
                              "type": "number"
                            },
                            "gamesPlayed": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "rating",
                            "gamesPlayed"
                          ],
                          "additionalProperties": false
                        },
                        "expectedScore": {
                          "type": "number"
                        },
                        "actualScore": {
                          "anyOf": [
                            {
                              "type": "number",
                              "enum": [
                                1
                              ]
                            },
                            {
                              "type": "number",
                              "enum": [
                                0.5
                              ]
                            },
                            {
                              "type": "number",
                              "enum": [
                                0
                              ]
                            }
                          ]
                        },
                        "delta": {
                          "type": "number"
                        },
                        "wasProvisional": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "playerId",
                        "before",
                        "after",
                        "expectedScore",
                        "actualScore",
                        "delta",
                        "wasProvisional"
                      ],
                      "additionalProperties": false
                    },
                    "upset": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "home",
                    "away",
                    "upset"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/matches/{id}": {
      "get": {
        "operationId": "getMatch",
        "summary": "Get a single match and its adjustment history",
        "tags": [
          "matches"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "match": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "sequence": {
                          "type": "number"
                        },
                        "homePlayerId": {
                          "type": "string"
                        },
                        "awayPlayerId": {
                          "type": "string"
                        },
                        "homeScore": {
                          "type": "number"
                        },
                        "awayScore": {
                          "type": "number"
                        },
                        "isVoid": {
                          "type": "boolean"
                        },
                        "playedAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        },
                        "sessionId": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "recordedBy": {
                          "type": "string"
                        },
                        "recordedAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        },
                        "decidedOnPenalties": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "sequence",
                        "homePlayerId",
                        "awayPlayerId",
                        "homeScore",
                        "awayScore",
                        "isVoid",
                        "playedAt",
                        "sessionId",
                        "recordedBy",
                        "recordedAt",
                        "decidedOnPenalties"
                      ],
                      "additionalProperties": false
                    },
                    "adjustments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "matchId": {
                            "type": "string"
                          },
                          "sequence": {
                            "type": "number"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "void",
                              "correct"
                            ]
                          },
                          "reason": {
                            "type": "string"
                          },
                          "newHomePlayerId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "newAwayPlayerId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "newHomeScore": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "newAwayScore": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "adjustedBy": {
                            "type": "string"
                          },
                          "adjustedAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          }
                        },
                        "required": [
                          "id",
                          "matchId",
                          "sequence",
                          "type",
                          "reason",
                          "newHomePlayerId",
                          "newAwayPlayerId",
                          "newHomeScore",
                          "newAwayScore",
                          "adjustedBy",
                          "adjustedAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "outcome": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "home": {
                              "type": "object",
                              "properties": {
                                "playerId": {
                                  "type": "string"
                                },
                                "before": {
                                  "type": "object",
                                  "properties": {
                                    "rating": {
                                      "type": "number"
                                    },
                                    "gamesPlayed": {
                                      "type": "number"
                                    }
                                  },
                                  "required": [
                                    "rating",
                                    "gamesPlayed"
                                  ],
                                  "additionalProperties": false
                                },
                                "after": {
                                  "type": "object",
                                  "properties": {
                                    "rating": {
                                      "type": "number"
                                    },
                                    "gamesPlayed": {
                                      "type": "number"
                                    }
                                  },
                                  "required": [
                                    "rating",
                                    "gamesPlayed"
                                  ],
                                  "additionalProperties": false
                                },
                                "expectedScore": {
                                  "type": "number"
                                },
                                "actualScore": {
                                  "anyOf": [
                                    {
                                      "type": "number",
                                      "enum": [
                                        1
                                      ]
                                    },
                                    {
                                      "type": "number",
                                      "enum": [
                                        0.5
                                      ]
                                    },
                                    {
                                      "type": "number",
                                      "enum": [
                                        0
                                      ]
                                    }
                                  ]
                                },
                                "delta": {
                                  "type": "number"
                                },
                                "wasProvisional": {
                                  "type": "boolean"
                                }
                              },
                              "required": [
                                "playerId",
                                "before",
                                "after",
                                "expectedScore",
                                "actualScore",
                                "delta",
                                "wasProvisional"
                              ],
                              "additionalProperties": false
                            },
                            "away": {
                              "type": "object",
                              "properties": {
                                "playerId": {
                                  "type": "string"
                                },
                                "before": {
                                  "type": "object",
                                  "properties": {
                                    "rating": {
                                      "type": "number"
                                    },
                                    "gamesPlayed": {
                                      "type": "number"
                                    }
                                  },
                                  "required": [
                                    "rating",
                                    "gamesPlayed"
                                  ],
                                  "additionalProperties": false
                                },
                                "after": {
                                  "type": "object",
                                  "properties": {
                                    "rating": {
                                      "type": "number"
                                    },
                                    "gamesPlayed": {
                                      "type": "number"
                                    }
                                  },
                                  "required": [
                                    "rating",
                                    "gamesPlayed"
                                  ],
                                  "additionalProperties": false
                                },
                                "expectedScore": {
                                  "type": "number"
                                },
                                "actualScore": {
                                  "anyOf": [
                                    {
                                      "type": "number",
                                      "enum": [
                                        1
                                      ]
                                    },
                                    {
                                      "type": "number",
                                      "enum": [
                                        0.5
                                      ]
                                    },
                                    {
                                      "type": "number",
                                      "enum": [
                                        0
                                      ]
                                    }
                                  ]
                                },
                                "delta": {
                                  "type": "number"
                                },
                                "wasProvisional": {
                                  "type": "boolean"
                                }
                              },
                              "required": [
                                "playerId",
                                "before",
                                "after",
                                "expectedScore",
                                "actualScore",
                                "delta",
                                "wasProvisional"
                              ],
                              "additionalProperties": false
                            },
                            "upset": {
                              "type": "boolean"
                            }
                          },
                          "required": [
                            "home",
                            "away",
                            "upset"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "match",
                    "adjustments",
                    "outcome"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/matches/{id}/void-preview": {
      "get": {
        "operationId": "previewVoidMatch",
        "summary": "Preview the ratings/ranks that would result from voiding a match, without voiding it (requires admin role)",
        "tags": [
          "matches"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "matchId": {
                      "type": "string"
                    },
                    "players": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "playerId": {
                            "type": "string"
                          },
                          "ratingBefore": {
                            "type": "number"
                          },
                          "gamesPlayedBefore": {
                            "type": "number"
                          },
                          "ratingAfter": {
                            "type": "number"
                          },
                          "gamesPlayedAfter": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "playerId",
                          "ratingBefore",
                          "gamesPlayedBefore",
                          "ratingAfter",
                          "gamesPlayedAfter"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "rankChanges": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "playerId": {
                            "type": "string"
                          },
                          "from": {
                            "type": "number"
                          },
                          "to": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "playerId",
                          "from",
                          "to"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "matchId",
                    "players",
                    "rankChanges"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [],
...
}